Renames RecurringPayment component, adds fontFamily to theme typography
